caption/requirements.txt
transformers==4.48.3
tqdm==4.67.1
pillow==11.1.0